+++ src/stream_wrap.cc
@@ -1,4 +1,7 @@
 #include "stream_wrap.h"
+#include "stream_base.h"
+#include "stream_base-inl.h"
+
 #include "env-inl.h"
 #include "env.h"
 #include "handle_wrap.h"
